# src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-commerce API",
    "version": "1.0.0",
    "description": "API documentation for the E-commerce platform."
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/orders": {
      "get": {
        "summary": "Get all orders",
        "tags": ["Orders"],
        "responses": {
          "200": {
            "description": "List of orders",
            "content": {
              "application/json": {
                "example": [
                  { "id": "6500c32f9d12345678c9ab44", "user": "64f1c72d8a1f3c4567d89ab1", "totalPrice": 2000, "orderStatus": "processing" }
                ]
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new order",
        "tags": ["Orders"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OrderCreate" },
              "example": {
                "user": "64f1c72d8a1f3c4567d89ab1",
                "orderItems": [{ "product": "6500b22f9d12345678c9ab33", "quantity": 2, "price": 1000 }],
                "shippingAddress": { "address": "123 Main St", "city": "New York", "postalCode": "10001", "country": "USA" },
                "paymentMethod": "paypal",
                "totalPrice": 2000
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Order created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Order" }, "example": { "id": "6500c32f9d12345678c9ab44", "orderStatus": "pending", "totalPrice": 2000 } } } },
          "400": { "description": "Bad request" }
        }
      }
    },
    "/orders/{id}": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
      "get": {
        "summary": "Get order by id",
        "tags": ["Orders"],
        "responses": { "200": { "description": "Order found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Order" } } } }, "404": { "description": "Order not found" } }
      },
      "put": {
        "summary": "Update an order (status/payment)",
        "tags": ["Orders"],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrderUpdate" }, "example": { "orderStatus": "shipped", "paymentStatus": "completed" } } } },
        "responses": { "200": { "description": "Order updated", "content": { "application/json": { "example": { "id": "6500c32f9d12345678c9ab44", "orderStatus": "shipped" } } } }, "404": { "description": "Order not found" } }
      },
      "delete": {
        "summary": "Delete an order",
        "tags": ["Orders"],
        "responses": { "200": { "description": "Order deleted", "content": { "application/json": { "example": { "message": "Order deleted" } } } }, "404": { "description": "Order not found" } }
      }
    },
    "/categories": {
      "get": {
        "summary": "Get all categories",
        "tags": ["Categories"],
        "responses": { "200": { "description": "List of categories", "content": { "application/json": { "example": [ { "id": "6500d42f9d12345678c9ab55", "name": "Electronics", "description": "Devices and gadgets" } ] } } } }
      },
      "post": {
        "summary": "Create a category",
        "tags": ["Categories"],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Category" }, "example": { "name": "Home", "description": "Household items" } } } },
        "responses": { "201": { "description": "Category created", "content": { "application/json": { "example": { "id": "6500d42f9d12345678c9ab55", "name": "Home" } } } } }
      }
    },
    "/categories/{id}": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
      "get": { "summary": "Get category by id", "tags": ["Categories"], "responses": { "200": { "description": "Category found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Category" } } } }, "404": { "description": "Category not found" } } },
      "put": { "summary": "Update category", "tags": ["Categories"], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Category" }, "example": { "name": "Electronics & Gadgets", "description": "Updated description" } } } }, "responses": { "200": { "description": "Category updated" }, "404": { "description": "Category not found" } } },
      "delete": { "summary": "Delete category", "tags": ["Categories"], "responses": { "200": { "description": "Category deleted" }, "404": { "description": "Category not found" } } }
    },
...
    "/cart/{userId}": {
      "parameters": [{ "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }],
      "get": {
        "summary": "Get a user's cart",
        "tags": ["Cart"],
        "responses": { "200": { "description": "User cart", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Cart" }, "example": { "user": "64f1c72d8a1f3c4567d89ab1", "items": [ { "product": "6500b22f9d12345678c9ab33", "quantity": 2 } ] } } } }, "404": { "description": "Cart not found" } }
      },
      "post": {
        "summary": "Add item to user's cart",
        "tags": ["Cart"],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CartAddItem" }, "example": { "product": "6500b22f9d12345678c9ab33", "quantity": 1 } } } },
        "responses": { "201": { "description": "Item added", "content": { "application/json": { "example": { "user": "64f1c72d8a1f3c4567d89ab1", "items": [ { "product": "6500b22f9d12345678c9ab33", "quantity": 1 } ] } } } } }
      },
      "put": {
        "summary": "Update item quantity in user's cart",
        "tags": ["Cart"],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CartAddItem" }, "example": { "product": "6500b22f9d12345678c9ab33", "quantity": 3 } } } },
        "responses": { "200": { "description": "Cart updated", "content": { "application/json": { "example": { "user": "64f1c72d8a1f3c4567d89ab1", "items": [ { "product": "6500b22f9d12345678c9ab33", "quantity": 3 } ] } } } } }
      },
      "delete": {
        "summary": "Clear entire user's cart",
        "tags": ["Cart"],
        "responses": { "200": { "description": "Cart cleared", "content": { "application/json": { "example": { "message": "Cart cleared" } } } }, "404": { "description": "Cart not found" } }
      }
    },
    "/cart/{userId}/{productId}": {
      "parameters": [
        { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } },
        { "name": "productId", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "delete": {
        "summary": "Remove a single product from user's cart",
        "tags": ["Cart"],
        "responses": { "200": { "description": "Product removed", "content": { "application/json": { "example": { "message": "Product removed from cart" } } } }, "404": { "description": "Not found" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
...
      "OrderCreate": {
        "type": "object",
        "required": ["user", "orderItems", "shippingAddress", "paymentMethod", "totalPrice"],
        "properties": {
          "user": { "type": "string" },
          "orderItems": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["product", "quantity", "price"],
              "properties": {
                "product": { "type": "string" },
                "quantity": { "type": "integer" },
                "price": { "type": "number" }
              }
            }
          },
          "shippingAddress": {
            "type": "object",
            "required": ["address", "city", "postalCode", "country"],
            "properties": { "address": { "type": "string" }, "city": { "type": "string" }, "postalCode": { "type": "string" }, "country": { "type": "string" } }
          },
          "paymentMethod": { "type": "string" },
          "paymentStatus": { "type": "string" },
          "orderStatus": { "type": "string" },
          "totalPrice": { "type": "number" }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "user": { "type": "string" },
          "orderItems": { "type": "array", "items": { "type": "object", "properties": { "product": { "type": "string" }, "quantity": { "type": "integer" }, "price": { "type": "number" } } } },
          "shippingAddress": { "type": "object", "properties": { "address": { "type": "string" }, "city": { "type": "string" }, "postalCode": { "type": "string" }, "country": { "type": "string" } } },
          "paymentMethod": { "type": "string" },
          "paymentStatus": { "type": "string" },
          "orderStatus": { "type": "string" },
          "totalPrice": { "type": "number" }
        }
      },
      "OrderUpdate": {
        "type": "object",
        "properties": { "orderStatus": { "type": "string" }, "paymentStatus": { "type": "string" } }
      },
      "Category": {
        "type": "object",
        "required": ["name"],
        "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "description": { "type": "string" } }
      },
      "Cart": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "user": { "type": "string" },
          "items": { "type": "array", "items": { "type": "object", "properties": { "product": { "type": "string" }, "quantity": { "type": "integer" } } } }
        }
      },
      "CartAddItem": {
        "type": "object",
        "required": ["product", "quantity"],
        "properties": { "product": { "type": "string" }, "quantity": { "type": "integer", "minimum": 1 } }
      },
...
    }
  }
}